+++ lib/src/actions/actions.dart
@@ -2,6 +2,7 @@ library;
 
 import 'package:collection/collection.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:layrz_logging/layrz_logging.dart';
 import 'package:layrz_models/layrz_models.dart';
 
 part 'actions.freezed.dart';
@@ -12,3 +13,4 @@ part 'src/action_subtype.dart';
 part 'src/action_property.dart';
 part 'src/action.dart';
 part 'src/action_geofence_settings.dart';
+part 'src/action_zigbee_settings.dart';